Fix build

+++ src/test-file.kt
@@ -4,5 +4,5 @@ import kotlin.math.*;
 
 fun main(args: Array<String>) {
     println(getSomething())
-    println(acos(1))
+    println(acos(1.toDouble()))
 }
